requirements.txt
fastapi==0.115.6
fastapi[standard]
pytest==8.3.4
httpx==0.28.1
pyinstrument==5.0.0
pyyaml==6.0.2